small change in AuthService.hpp

+++ include/domain/entities/Cat.hpp
@@ -3,21 +3,21 @@
 
 class Cat : public Pet {
 private:
-  string furLength; // Độ dài lông: Short, Medium, Long
+    string furLength; // Độ dài lông: Short, Medium, Long
 
 public:
-  Cat(int id, const string &name, const string &breed, int age, double price,
-      const string &furLength = "Short")
-      : Pet(id, name, breed, age, price), furLength(furLength) {}
+    Cat(int id, const string &name, const string &breed, int age, float price,
+        const string &furLength = "Short")
+        : Pet(id, name, breed, age, price), furLength(furLength) {}
 
-  string getFurLength() const { return furLength; }
-  void setFurLength(const string &value) { furLength = value; }
+    string getFurLength() const { return furLength; }
+    void setFurLength(const string &value) { furLength = value; }
 
-  string getType() const override { return "Cat"; }
+    string getType() const override { return "Cat"; }
 
-  void showInfo() const override {
-    cout << "=== CAT INFO ===\n";
-    Pet::showInfo();
-    cout << "Fur Length: " << furLength << "\n";
-  }
+    void showInfo() const override {
+        cout << "=== CAT INFO ===\n";
+        Pet::showInfo();
+        cout << "Fur Length: " << furLength << "\n";
+    }
 };
